Move projects to their own module file
