+++ public/src/modules/database.js
@@ -13,4 +13,17 @@ const add = async (collection, data) => {
   return result;
 };
 
+const getUserData = async () => {
+  let collRef = firestore.collection(collection);
+  let result;
+
+  try {
+    result = await collRef.add(data);
+  }catch(error) {
+    result = await error;
+  };
+
+  return result;
+}
+
 export {add};
